requirements.txt
requests
pandas
seaborn
matplotlib
plotly
numpy
biopython
kaleido